More refactoring of tests.
