+++ pcgrandom/test/test_pcg_xsh_rr_v0.py
@@ -3,7 +3,6 @@ Tests for the PCG_XSH_RR_V0 generator.
 """
 from __future__ import division
 
-import math
 import unittest
 
 from pcgrandom.pcg_xsh_rr_v0 import PCG_XSH_RR_V0
@@ -40,19 +39,10 @@ seq0_seed12345_shuffle = [
 
 
 class Test_PCG_XSH_RR_V0(TestCommon, unittest.TestCase):
+    gen_class = PCG_XSH_RR_V0
+
     def setUp(self):
-        self.gen = PCG_XSH_RR_V0(seed=15206, sequence=1729)
-
-    def test_creation_without_seed(self):
-        gen1 = PCG_XSH_RR_V0()
-        gen2 = PCG_XSH_RR_V0()
-
-        sample1 = [gen1.random() for _ in range(10)]
-        sample2 = [gen2.random() for _ in range(10)]
-
-        # Possible in theory for these two samples to be identical; vanishingly
-        # unlikely in practice.
-        self.assertNotEqual(sample1, sample2)
+        self.gen = self.gen_class(seed=15206, sequence=1729)
 
     def test_reproducibility(self):
         gen = PCG_XSH_RR_V0(seed=12345, sequence=0)
@@ -86,62 +76,3 @@ class Test_PCG_XSH_RR_V0(TestCommon, unittest.TestCase):
         population = list(range(20))
         gen.shuffle(population)
         self.assertEqual(population, seq0_seed12345_shuffle)
-
-    def test_sequence_default(self):
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345)
-        self.assertEqual(gen1.getstate(), gen2.getstate())
-
-    def test_independent_sequences(self):
-        # Crude statistical test for lack of correlation. If X and Y are
-        # independent and uniform on [0, 1], then (X - 0.5) * (Y - 0.5) has
-        # mean 0 and standard deviation 1/12. By the central limit theorem, we
-        # expect the average V of N such independent values to be roughly
-        # normally distributed with mean 0 and standard deviation 1 /
-        # (12*sqrt(N)). So we expect |V| to be at most 1 / (4*sqrt(N)) with
-        # over 99% probability.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
-        N = 10000
-        xs = [gen1.random() for _ in range(N)]
-        ys = [gen2.random() for _ in range(N)]
-        v = sum((x - 0.5) * (y - 0.5) for x, y in zip(xs, ys)) / N
-        # Check we're within 3 standard deviations of the mean.
-        self.assertLess(abs(v), 0.25/math.sqrt(N))
-
-    def test_no_shared_state(self):
-        # Get samples first from gen1, then from gen2.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
-        sample1_1 = [gen1.random() for _ in range(10)]
-        sample2_1 = [gen2.random() for _ in range(10)]
-
-        # Now in the opposite order: from gen2, then from gen1.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
-        sample2_2 = [gen2.random() for _ in range(10)]
-        sample1_2 = [gen1.random() for _ in range(10)]
-
-        # Now interleaved.
-        gen1 = PCG_XSH_RR_V0(seed=12345, sequence=0)
-        gen2 = PCG_XSH_RR_V0(seed=12345, sequence=1)
-        sample1_3 = []
-        sample2_3 = []
-        for _ in range(10):
-            sample1_3.append(gen1.random())
-            sample2_3.append(gen2.random())
-
-        # Results should be the same in all cases.
-        self.assertEqual(sample1_1, sample1_2)
-        self.assertEqual(sample1_1, sample1_3)
-        self.assertEqual(sample2_1, sample2_2)
-        self.assertEqual(sample2_1, sample2_3)
-
-    def test_restore_state_from_different_generator(self):
-        gen = PCG_XSH_RR_V0(seed=15206, sequence=27)
-        state = gen.getstate()
-
-        bad_version = 'pcgrandom.PCG_XSH_RR_V1'
-        bad_state = (bad_version,) + state[1:]
-        with self.assertRaises(ValueError):
-            gen.setstate(bad_state)
